Viewer: Improved layout engine
We now merge staves that are close together or even overlap. This results in less visual artifacts
between staves, especially when they are tightly packed vertically in the original material.

Overlapping staff annotations are now handled perfectly fine, it is even recommended to make liberal
use of this for best looking results. (Previously, the automatic recognition in the editor even
automatically clamped any overlap between staves)